# Garbage collection speed optimization