In our docker registries current GC implementation became unacceptably slow, so I have prepared some improvements to reduce computational complexity of the process.

GC is functionally the same, I compared current implementation's dry run output with new version's and it looks similar (there were some differences due to docker repositories changing as I worked on this).

Speed gain may vary depending on repository structure and storage backend used.
In one of our repositories (~500 repos; ~22k tags, ~500k blobs, swift driver with #2642 applied) GC dry run time dropped from ~75h to less than 9h (8.5 times faster).

I have been running dry runs only, so I do not recommend to merge this in for now - I'm leaving this code for review. I'll let you know the results of running this on our registries.

First registry: only 17 repositories, 104 tags, 500k blobs due to frequent image rebuilding and tag creation/deletion (automatic tests):
```
840 blobs marked, 505651 blobs and 52 manifests eligible for deletion
GC mark phase: 45s
GC manifest removal phase 21s
GC blob scan phase 4h 21m 8s
GC blob removal phase 14h 42m 54s
```
Next GC run took less than one minute.

Second registry: 331 repositories, 1450 tags, ~62k blobs:
```
11209 blobs marked, 51540 blobs and 5954 manifests eligible for deletion
GC mark phase: 39m 22s
GC manifest removal phase: 34m 15s
GC blob scan phase: 28m 10s
GC blob removal phase: 4h 2m 18s
```