@bartebor 2018-07-19
Looks like tests got killed (timeout?). How to run them again?