@bartebor 2024-05-28
Thanks @milosgajdos, I'm not interested in this issue anymore. Our  installations has been working for all these years with this PR merged and no data was lost :) Anyway, we are replacing distribution with quay due to increased requirements and I have no time to continue with this PR. If someone is interetsed, feel free to take over.